@@ -28,12 +28,14 @@
   },
   "homepage": "https://github.com/vkbansal/react-contextmenu",
   "dependencies": {
+    "autobind-decorator": "^1.3.2",
     "classnames": "^2.1.3",
     "flummox": "^3.5.2",
     "invariant": "^2.1.0",
     "lodash.isobject": "^3.0.2",
-    "object-assign": "^3.0.0",
-    "react": "^0.13.3"
+    "object-assign": "^4.0.1",
+    "react": "^0.14.0",
+    "react-dom": "^0.14.0"
   },
   "devDependencies": {
     "babel": "^5.8.21",
@@ -43,7 +45,7 @@
     "eslint": "^1.1.0",
     "eslint-plugin-react": "^3.2.2",
     "node-libs-browser": "^0.5.2",
-    "react-router": "^0.13.3",
+    "react-router": "^1.0.0-rc3",
     "webpack": "^1.11.0",
     "webpack-dev-server": "^1.10.1"
   }
